major update
Rename objects to follow naming convention
Add a "Bring To Front" option in the system tray context menu
Add a toggle for "Always on Top" to the system tray context menu
Improve the layout of controls
Improve the visual style of combo boxes
Fix autocomplete for hotkey & target comboboxes
Apply code formatting preferences
Improve tab stop indexes
Consolidate always on top settings into a property
Add a bring to front event handler
Disable toast setting members when the toast is disabled.

+++ ToastForm/Extensions/FloatingPointExtensions.cs
@@ -8,7 +8,7 @@
         }
         public static bool EqualsWithin(this float f, double compare, double epsilon = double.Epsilon)
         {
-            return (((double)f) - compare <= epsilon);
+            return (f - compare <= epsilon);
         }
         public static bool EqualsWithin(this double d, double compare, double epsilon = double.Epsilon)
         {
